+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## v1.0.2
+
+- chore: optimize cmd/server/main.go
+- chore: optimize cmd/consumer/main.go
+
 ## v1.0.1
 
 - chore: move main.go to cmd/server